adds tests for gallery module
