+++ test/unit/gallery-store-config.spec.js
@@ -1,0 +1,8 @@
+// gallery-store-config.spec.js
+
+import { galleryState, galleryMutations } from "../../src/store/gallery"
+
+export default {
+  state: galleryState,
+  mutations: galleryMutations,
+}
