+++ test/unit/jest.conf.js
@@ -14,7 +14,7 @@ module.exports = {
   snapshotSerializers: ["<rootDir>/node_modules/jest-serializer-vue"],
   setupFiles: ["<rootDir>/test/unit/setup"],
   coverageDirectory: "<rootDir>/test/unit/coverage",
-  testPathIgnorePatterns: ["/node_modules/", "/cypress/"],
+  testPathIgnorePatterns: ["/node_modules/", "/cypress/", "<rootDir>/test/unit/gallery-store-config.spec.js"],
   collectCoverageFrom: [
     "<rootDir>/src/ExampleComponent.js",
     "<rootDir>/src/elements/*.{js,vue}",
@@ -23,6 +23,9 @@ module.exports = {
     "<rootDir>/docs/**/*.{js,vue}",
     "!<rootDir>/cypress/**",
     "!<rootDir>/src/main.js",
+    "!<rootDir>/test/unit/gallery-store-config.spec.js",
+    "!<rootDir>/src/patterns/_OrderManager*",
+    "!<rootDir>/src/patterns/OrderManager*",
     "!<rootDir>/src/router/index.js",
     "!<rootDir>/node_modules/**",
     "!<rootDir>/src/system.js",
